pyproject.toml
[tool.poetry]
name = "jobs-bundle"
version = "0.3.1"
description = "Databricks jobs management bundle for the Daipe Framework"
readme = "README.md"
repository = "https://github.com/daipe-ai/jobs-bundle"

license = "MIT"

authors = [
    "Jiri Koutny <jiri.koutny@datasentics.com>"
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

include = ["*.yaml"]

packages = [
    { include = "jobsbundle", from = "src" },
]

[tool.poetry.dependencies]
python-box = "^3.4"
python = "^3.7"
pyfony-bundles = "0.4.*"
console-bundle = "^0.5"
databricks-cli = ">=0.16.0"
pygit2 = "^1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.0"
black = "^20.8b1"
flake8-fixme = "^1.1.1"
pep8-naming = "^0.11.1"
poethepoet = "^0.10.0"
pyfony-core = "0.8.*"

[tool.poetry.plugins."pyfony.bundle"]
create = "jobsbundle.JobsBundle:JobsBundle"

[pyfony.bootstrap]
container_init = "pyfonycore.container.container_init:init_with_current_bundle"
root_module_name = "jobsbundle"

[tool.black]
line-length = 140

[tool.poe.tasks]
black = "black src"
black-check = "black --check src"
flake8 = "flake8 --ignore=E501,W503 src"
